Fixes as per review
