+++ src/Stratis.Bitcoin.IntegrationTests/EnvironmentMockUpHelpers/NodeRunners.cs
@@ -1,12 +1,9 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using Microsoft.Extensions.DependencyInjection;
 using NBitcoin;
 using NBitcoin.Protocol;
 using Stratis.Bitcoin.Builder;
-using Stratis.Bitcoin.Builder.Feature;
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Features.Api;
 using Stratis.Bitcoin.Features.BlockStore;
@@ -15,8 +12,6 @@ using Stratis.Bitcoin.Features.MemoryPool;
 using Stratis.Bitcoin.Features.Miner;
 using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Features.Wallet;
-using Stratis.Bitcoin.Interfaces;
-using Stratis.Bitcoin.Utilities;
 
 namespace Stratis.Bitcoin.IntegrationTests.EnvironmentMockUpHelpers
 {
@@ -331,56 +326,6 @@ namespace Stratis.Bitcoin.IntegrationTests.EnvironmentMockUpHelpers
             }
 
             return node;
-        }
-    }
-
-    public static class FullNodeTestBuilderExtension
-    {
-        /// <summary>
-        /// Substitute the <see cref="IDateTimeProvider"/> for a given feature.
-        /// </summary>
-        /// <typeparam name="T">The feature to substitute the provider for.</typeparam>
-        public static IFullNodeBuilder SubstituteDateTimeProviderFor<T>(this IFullNodeBuilder fullNodeBuilder, IDateTimeProvider provider)
-        {
-            fullNodeBuilder.ConfigureFeature(features =>
-            {
-                var feature = features.FeatureRegistrations.FirstOrDefault(f => f.FeatureType == typeof(T));
-                if (feature != null)
-                {
-                    feature.FeatureServices(services =>
-                    {
-                        ServiceDescriptor service = services.FirstOrDefault(s => s.ServiceType == typeof(IDateTimeProvider));
-                        if (service != null)
-                            services.Remove(service);
-                        services.AddSingleton(provider);
-                    });
-                }
-            });
-
-            return fullNodeBuilder;
-        }
-
-        public static IFullNodeBuilder MockIBD(this IFullNodeBuilder fullNodeBuilder)
-        {
-            fullNodeBuilder.ConfigureFeature(features =>
-            {
-                foreach (IFeatureRegistration feature in features.FeatureRegistrations)
-                {
-                    feature.FeatureServices(services =>
-                    {
-                        // Get default IBD implementation and replace it with the mock.
-                        ServiceDescriptor ibdService = services.FirstOrDefault(x => x.ServiceType == typeof(IInitialBlockDownloadState));
-
-                        if (ibdService != null)
-                        {
-                            services.Remove(ibdService);
-                            services.AddSingleton<IInitialBlockDownloadState, InitialBlockDownloadStateMock>();
-                        }
-                    });
-                }
-            });
-
-            return fullNodeBuilder;
         }
     }
 }
